Remove sparse solver and migrate to ifx
This commit removes the old Berkeley sparse solver that we have not used
for a long time. The sparse solver does not get compiled nicely with
newer Intel compilers because of antiquated C syntaxes. Thus, we
decided to remove it.
At the same time, some type casts are added to appease more stringent
C++ rules for the new icpx.
Note that one of dependencies, HEClib library (ver 20216) codes got some
updates for ifx as well.
No functional changes are introduced, and this version can be built with
previous Intel compilers, e.g. ifort 2022.1.

DSM2-876, DSM2-901

+++ dsm2/src/hdf_tidefile/dsm2_tidefile_input_storage_hydro_comp_point.cpp
@@ -155,9 +155,9 @@ TableDescription hydro_comp_point_table_description(){
                };
 
   const size_t foffsets[] ={
-             ((char*)&default_struct.comp_index - (char*)&default_struct),
-             ((char*)&default_struct.channel - (char*)&default_struct),
-             ((char*)&default_struct.distance - (char*)&default_struct)
+             (size_t)((char*)&default_struct.comp_index - (char*)&default_struct),
+             (size_t)((char*)&default_struct.channel - (char*)&default_struct),
+             (size_t)((char*)&default_struct.distance - (char*)&default_struct)
                            };
 
   const size_t fsizes[] = {
