+++ dsm2/src/hdf_tidefile/dsm2_tidefile_input_storage_node_flow_connections.cpp
@@ -236,13 +236,13 @@ TableDescription node_flow_connections_table_description(){
                };
 
   const size_t foffsets[] ={
-             ((char*)&default_struct.connection_index - (char*)&default_struct),
-             ((char*)&default_struct.int_node_no - (char*)&default_struct),
-             ((char*)&default_struct.ext_node_no - (char*)&default_struct),
-             ((char*)&default_struct.node_flow_index - (char*)&default_struct),
-             ((char*)&default_struct.flow_index - (char*)&default_struct),
-             ((char*)&default_struct.flow_name - (char*)&default_struct),
-             ((char*)&default_struct.flow_type - (char*)&default_struct)
+             (size_t)((char*)&default_struct.connection_index - (char*)&default_struct),
+             (size_t)((char*)&default_struct.int_node_no - (char*)&default_struct),
+             (size_t)((char*)&default_struct.ext_node_no - (char*)&default_struct),
+             (size_t)((char*)&default_struct.node_flow_index - (char*)&default_struct),
+             (size_t)((char*)&default_struct.flow_index - (char*)&default_struct),
+             (size_t)((char*)&default_struct.flow_name - (char*)&default_struct),
+             (size_t)((char*)&default_struct.flow_type - (char*)&default_struct)
                            };
 
   const size_t fsizes[] = {
